+++ test/unittests/__init__.py
@@ -13,4 +13,6 @@ from url        import *
 
 from job.description import *
 from job.issues      import *
-from job.misc        import * 
+from job.misc        import *
+
+from file.directory  import *